Fix review requested changes
