+++ cypress.config.js
@@ -4,7 +4,6 @@ module.exports = defineConfig({
   e2e: {
     baseUrl: 'http://localhost:3000',
     supportFile: false
-    },
-    fixturesFolder: false,
+    }
   },
-})
+)
