--- cypress/support/commands.js
@@ -1,1 +1,0 @@
-import '@testing-library/cypress/add-commands'
