--- cypress/support/e2e.js
@@ -1,1 +1,0 @@
-import './commands'
